frontend/requirements.txt
alembic
autoenv
blinker
certifi
cffi
chardet
click
cryptography
dnspython
dominate
email-validator
Flask
Flask-Bootstrap
Flask-DotEnv
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-Uploads
Flask-WTF
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
marshmallow
passlib
protobuf
pycparser
PyMySQL
python-dateutil
python-dotenv
python-editor
requests
six
SQLAlchemy
urllib3
visitor
Werkzeug
WTForms